live-comments/live_comments_be/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
bidict==0.22.1
billiard==4.1.0
celery==5.3.4
celery-batches==0.8.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.6
django-cors-headers==4.3.0
django-redis==5.4.0
djangorestframework==3.14.0
flower==2.0.1
h11==0.14.0
humanize==4.8.0
kombu==5.3.2
prometheus-client==0.17.1
prompt-toolkit==3.0.39
python-dateutil==2.8.2
python-engineio==4.8.0
python-socketio==5.10.0
pytz==2023.3.post1
redis==5.0.1
simple-websocket==1.0.0
six==1.16.0
sqlparse==0.4.4
structlog==23.2.0
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.8
wsproto==1.2.0
